refactor: moved markdown commands
